requirements.txt
requests==2.31.0
python-dotenv==1.0.0
slack-sdk==3.26.1
pydantic==2.5.3
flask==2.3.3 